Switch to using x11rb for event handling
This switches the event handling code to use a more decentralized
organizational structure to avoid needing to have the huge 2000 line
long match statement that every other backend has. In fact, the code can
even be split up among multiple files.

Unfortunately this commit breaks XIM, which I will fix before I merge
the PR.

+++ src/platform_impl/linux/x11/dnd.rs
@@ -44,7 +44,7 @@ impl From<io::Error> for DndDataParseError {
 pub(crate) struct Dnd {
     xconn: Arc<XConnection>,
     // Populated by XdndEnter event handler
-    pub version: Option<c_long>,
+    pub version: Option<u32>,
     pub type_list: Option<Vec<xproto::Atom>>,
     // Populated by XdndPosition event handler
     pub source_window: Option<xproto::Window>,
